+++ docs/sphinx/changelog.rst
@@ -32,3 +32,4 @@ New Features
 - Added the decomposition of a sphere for the purposes of integration and contact detection. (:merge:`10`). `Nathan Miller`_.
 - Added the capability to integrate 2D quadratic elements and surface meshes composed of quadratic elements. (:merge:`11`). `Nathan Miller`_.
 - Added a more general form of the distance calculation to support random deformations. (:merge:`12`). `Nathan Miller`_.
+- Initial commit of the asp base class with a linear elastic local particle energy definition. (:merge:`14`). `Nathan Miller`_.